refactor: modularize associated iterators
