+++ src/lib.rs
@@ -65,7 +65,17 @@ pub extern crate alloc;
 #[cfg(any(test, feature = "std"))]
 extern crate std;
 
+mod drain;
+mod extract_if;
+mod into_iter;
 mod rawsmallvec;
+mod splice;
+
+pub use drain::Drain;
+pub use extract_if::ExtractIf;
+pub use into_iter::IntoIter;
+pub use splice::Splice;
+
 #[cfg(test)]
 mod tests;
 
@@ -353,484 +363,6 @@ impl<T, const N: usize> Default for SmallVec<T, N> {
         Self::new()
     }
 }
-
-/// An iterator that removes the items from a `SmallVec` and yields them by
-/// value.
-///
-/// Returned from [`SmallVec::drain`][1].
-///
-/// [1]: struct.SmallVec.html#method.drain
-pub struct Drain<'a, T: 'a, const N: usize> {
-    // `vec` points to a valid object within its lifetime.
-    // This is ensured by the fact that we're holding an iterator to its items.
-    //
-    // # Safety
-    //
-    // Members in vec[tail_start..tail_start + tail_len] are initialized
-    // even though vec has length < tail_start
-    tail_start: usize,
-    tail_len: usize,
-    iter: core::slice::Iter<'a, T>,
-    vec: core::ptr::NonNull<SmallVec<T, N>>,
-}
-
-impl<'a, T: 'a, const N: usize> Iterator for Drain<'a, T, N> {
-    type Item = T;
-
-    #[inline]
-    fn next(&mut self) -> Option<T> {
-        // SAFETY: we shrunk the length of the vector so it no longer owns these items,
-        // and we can take ownership of them.
-        self.iter
-            .next()
-            .map(|reference| unsafe { core::ptr::read(reference) })
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.iter.size_hint()
-    }
-}
-
-impl<'a, T: 'a, const N: usize> DoubleEndedIterator for Drain<'a, T, N> {
-    #[inline]
-    fn next_back(&mut self) -> Option<T> {
-        // SAFETY: see above
-        self.iter
-            .next_back()
-            .map(|reference| unsafe { core::ptr::read(reference) })
-    }
-}
-
-impl<T, const N: usize> ExactSizeIterator for Drain<'_, T, N> {
-    #[inline]
-    fn len(&self) -> usize {
-        self.iter.len()
-    }
-}
-
-impl<T, const N: usize> core::iter::FusedIterator for Drain<'_, T, N> {}
-
-impl<'a, T: 'a, const N: usize> Drop for Drain<'a, T, N> {
-    fn drop(&mut self) {
-        /// Moves back the un-`Drain`ed elements to restore the original `Vec`.
-        struct DropGuard<'r, 'a, T, const N: usize>(&'r mut Drain<'a, T, N>);
-
-        impl<'r, 'a, T, const N: usize> Drop for DropGuard<'r, 'a, T, N> {
-            fn drop(&mut self) {
-                if self.0.tail_len > 0 {
-                    unsafe {
-                        let source_vec = self.0.vec.as_mut();
-                        // memmove back untouched tail, update to new length
-                        let start = source_vec.len();
-                        let tail = self.0.tail_start;
-                        if tail != start {
-                            let ptr = source_vec.as_mut_ptr();
-                            let src = ptr.add(tail);
-                            let dst = ptr.add(start);
-                            core::ptr::copy(src, dst, self.0.tail_len);
-                        }
-                        source_vec.set_len(start + self.0.tail_len);
-                    }
-                }
-            }
-        }
-
-        let iter = core::mem::take(&mut self.iter);
-        let drop_len = iter.len();
-
-        let mut vec = self.vec;
-
-        if SmallVec::<T, N>::IS_ZST {
-            // ZSTs have no identity, so we don't need to move them around, we only need to
-            // drop the correct amount. this can be achieved by manipulating the
-            // Vec length instead of moving values out from `iter`.
-            unsafe {
-                let vec = vec.as_mut();
-                let old_len = vec.len();
-                vec.set_len(old_len + drop_len + self.tail_len);
-                vec.truncate(old_len + self.tail_len);
-            }
-
-            return;
-        }
-
-        // ensure elements are moved back into their appropriate places, even when
-        // drop_in_place panics
-        let _guard = DropGuard(self);
-
-        if drop_len == 0 {
-            return;
-        }
-
-        // as_slice() must only be called when iter.len() is > 0 because
-        // it also gets touched by vec::Splice which may turn it into a dangling pointer
-        // which would make it and the vec pointer point to different allocations which
-        // would lead to invalid pointer arithmetic below.
-        let drop_ptr = iter.as_slice().as_ptr();
-
-        unsafe {
-            // drop_ptr comes from a slice::Iter which only gives us a &[T] but for
-            // drop_in_place a pointer with mutable provenance is necessary.
-            // Therefore we must reconstruct it from the original vec but also
-            // avoid creating a &mut to the front since that could invalidate
-            // raw pointers to it which some unsafe code might rely on.
-            let vec_ptr = vec.as_mut().as_mut_ptr();
-            // May be replaced with the line below later, once this crate's MSRV is >= 1.87.
-            //let drop_offset = drop_ptr.offset_from_unsigned(vec_ptr);
-            let drop_offset = drop_ptr.offset_from(vec_ptr) as usize;
-            let to_drop = core::ptr::slice_from_raw_parts_mut(vec_ptr.add(drop_offset), drop_len);
-            core::ptr::drop_in_place(to_drop);
-        }
-    }
-}
-
-impl<T, const N: usize> Drain<'_, T, N> {
-    #[must_use]
-    pub fn as_slice(&self) -> &[T] {
-        self.iter.as_slice()
-    }
-
-    /// The range from `self.vec.len` to `self.tail_start` contains elements
-    /// that have been moved out.
-    /// Fill that range as much as possible with new elements from the
-    /// `replace_with` iterator. Returns `true` if we filled the entire
-    /// range. (`replace_with.next()` didn’t return `None`.)
-    unsafe fn fill<I: Iterator<Item = T>>(&mut self, replace_with: &mut I) -> bool {
-        let vec = unsafe { self.vec.as_mut() };
-        let range_start = vec.len();
-        let range_end = self.tail_start;
-        let range_slice = unsafe {
-            core::slice::from_raw_parts_mut(
-                vec.as_mut_ptr().add(range_start),
-                range_end - range_start,
-            )
-        };
-
-        for place in range_slice {
-            if let Some(new_item) = replace_with.next() {
-                unsafe { core::ptr::write(place, new_item) };
-                vec.set_len(vec.len() + 1);
-            } else {
-                return false;
-            }
-        }
-        true
-    }
-
-    /// Makes room for inserting more elements before the tail.
-    #[track_caller]
-    unsafe fn move_tail(&mut self, additional: usize) {
-        let vec = unsafe { self.vec.as_mut() };
-        let len = self.tail_start + self.tail_len;
-
-        // Test
-        let old_len = vec.len();
-        vec.set_len(len);
-        vec.reserve(additional);
-        vec.set_len(old_len);
-
-        let new_tail_start = self.tail_start + additional;
-        unsafe {
-            let src = vec.as_ptr().add(self.tail_start);
-            let dst = vec.as_mut_ptr().add(new_tail_start);
-            core::ptr::copy(src, dst, self.tail_len);
-        }
-        self.tail_start = new_tail_start;
-    }
-}
-
-/// An iterator which uses a closure to determine if an element should be
-/// removed.
-///
-/// Returned from [`SmallVec::extract_if`][1].
-///
-/// [1]: struct.SmallVec.html#method.extract_if
-pub struct ExtractIf<'a, T, const N: usize, F>
-where
-    F: FnMut(&mut T) -> bool,
-{
-    vec: &'a mut SmallVec<T, N>,
-    /// The index of the item that will be inspected by the next call to `next`.
-    idx: usize,
-    /// Elements at and beyond this point will be retained. Must be equal or
-    /// smaller than `old_len`.
-    end: usize,
-    /// The number of items that have been drained (removed) thus far.
-    del: usize,
-    /// The original length of `vec` prior to draining.
-    old_len: usize,
-    /// The filter test predicate.
-    pred: F,
-}
-
-impl<T, const N: usize, F> core::fmt::Debug for ExtractIf<'_, T, N, F>
-where
-    F: FnMut(&mut T) -> bool,
-    T: core::fmt::Debug,
-{
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.debug_tuple("ExtractIf")
-            .field(&self.vec.as_slice())
-            .finish()
-    }
-}
-
-impl<T, F, const N: usize> Iterator for ExtractIf<'_, T, N, F>
-where
-    F: FnMut(&mut T) -> bool,
-{
-    type Item = T;
-
-    fn next(&mut self) -> Option<T> {
-        unsafe {
-            while self.idx < self.end {
-                let i = self.idx;
-                let v = core::slice::from_raw_parts_mut(self.vec.as_mut_ptr(), self.old_len);
-                let drained = (self.pred)(&mut v[i]);
-                // Update the index *after* the predicate is called. If the index
-                // is updated prior and the predicate panics, the element at this
-                // index would be leaked.
-                self.idx += 1;
-                if drained {
-                    self.del += 1;
-                    return Some(core::ptr::read(&v[i]));
-                } else if self.del > 0 {
-                    let del = self.del;
-                    let src: *const T = &v[i];
-                    let dst: *mut T = &mut v[i - del];
-                    core::ptr::copy_nonoverlapping(src, dst, 1);
-                }
-            }
-            None
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (0, Some(self.end - self.idx))
-    }
-}
-
-impl<T, F, const N: usize> Drop for ExtractIf<'_, T, N, F>
-where
-    F: FnMut(&mut T) -> bool,
-{
-    fn drop(&mut self) {
-        unsafe {
-            if self.idx < self.old_len && self.del > 0 {
-                // This is a pretty messed up state, and there isn't really an
-                // obviously right thing to do. We don't want to keep trying
-                // to execute `pred`, so we just backshift all the unprocessed
-                // elements and tell the vec that they still exist. The backshift
-                // is required to prevent a double-drop of the last successfully
-                // drained item prior to a panic in the predicate.
-                let ptr = self.vec.as_mut_ptr();
-                let src = ptr.add(self.idx);
-                let dst = src.sub(self.del);
-                let tail_len = self.old_len - self.idx;
-                src.copy_to(dst, tail_len);
-            }
-            self.vec.set_len(self.old_len - self.del);
-        }
-    }
-}
-
-pub struct Splice<'a, I: Iterator + 'a, const N: usize> {
-    drain: Drain<'a, I::Item, N>,
-    replace_with: I,
-}
-
-impl<'a, I, const N: usize> core::fmt::Debug for Splice<'a, I, N>
-where
-    I: Debug + Iterator + 'a,
-    <I as Iterator>::Item: Debug,
-{
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.debug_tuple("Splice").field(&self.drain).finish()
-    }
-}
-
-impl<I: Iterator, const N: usize> Iterator for Splice<'_, I, N> {
-    type Item = I::Item;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.drain.next()
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.drain.size_hint()
-    }
-}
-
-impl<I: Iterator, const N: usize> DoubleEndedIterator for Splice<'_, I, N> {
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.drain.next_back()
-    }
-}
-
-impl<I: Iterator, const N: usize> ExactSizeIterator for Splice<'_, I, N> {}
-
-impl<I: Iterator, const N: usize> Drop for Splice<'_, I, N> {
-    fn drop(&mut self) {
-        self.drain.by_ref().for_each(drop);
-        // At this point draining is done and the only remaining tasks are splicing
-        // and moving things into the final place.
-        // Which means we can replace the slice::Iter with pointers that won't point to
-        // deallocated memory, so that Drain::drop is still allowed to call
-        // iter.len(), otherwise it would break the ptr.sub_ptr contract.
-        self.drain.iter = [].iter();
-
-        unsafe {
-            if self.drain.tail_len == 0 {
-                self.drain.vec.as_mut().extend(self.replace_with.by_ref());
-                return;
-            }
-
-            // First fill the range left by drain().
-            if !self.drain.fill(&mut self.replace_with) {
-                return;
-            }
-
-            // There may be more elements. Use the lower bound as an estimate.
-            // FIXME: Is the upper bound a better guess? Or something else?
-            let (lower_bound, _upper_bound) = self.replace_with.size_hint();
-            if lower_bound > 0 {
-                self.drain.move_tail(lower_bound);
-                if !self.drain.fill(&mut self.replace_with) {
-                    return;
-                }
-            }
-
-            // Collect any remaining elements.
-            let mut collected = self
-                .replace_with
-                .by_ref()
-                .collect::<SmallVec<I::Item, N>>()
-                .into_iter();
-            // Now we have an exact count.
-            if collected.len() > 0 {
-                self.drain.move_tail(collected.len());
-                let filled = self.drain.fill(&mut collected);
-                debug_assert!(filled);
-                debug_assert_eq!(collected.len(), 0);
-            }
-        }
-        // Let `Drain::drop` move the tail back if necessary and restore
-        // `vec.len`.
-    }
-}
-
-/// An iterator that consumes a `SmallVec` and yields its items by value.
-///
-/// Returned from [`SmallVec::into_iter`][1].
-///
-/// [1]: struct.SmallVec.html#method.into_iter
-pub struct IntoIter<T, const N: usize> {
-    // # Safety
-    //
-    // `end` decides whether the data lives on the heap or not
-    //
-    // The members from begin..end are initialized
-    raw: RawSmallVec<T, N>,
-    begin: usize,
-    end: TaggedLen<T>,
-    _marker: PhantomData<T>,
-}
-
-// SAFETY: IntoIter has unique ownership of its contents.  Sending (or sharing)
-// an `IntoIter<T, N>` is equivalent to sending (or sharing) a `SmallVec<T, N>`.
-unsafe impl<T, const N: usize> Send for IntoIter<T, N> where T: Send {}
-unsafe impl<T, const N: usize> Sync for IntoIter<T, N> where T: Sync {}
-
-impl<T, const N: usize> IntoIter<T, N> {
-    #[inline]
-    const fn as_ptr(&self) -> *const T {
-        let on_heap = self.end.on_heap();
-        if on_heap {
-            // SAFETY: vector is on the heap
-            unsafe { self.raw.as_ptr_heap() }
-        } else {
-            self.raw.as_ptr_inline()
-        }
-    }
-
-    #[inline]
-    const fn as_mut_ptr(&mut self) -> *mut T {
-        let on_heap = self.end.on_heap();
-        if on_heap {
-            // SAFETY: vector is on the heap
-            unsafe { self.raw.as_mut_ptr_heap() }
-        } else {
-            self.raw.as_mut_ptr_inline()
-        }
-    }
-
-    #[inline]
-    pub const fn as_slice(&self) -> &[T] {
-        // SAFETY: The members in self.begin..self.end.value() are all initialized
-        // So the pointer arithmetic is valid, and so is the construction of the slice
-        unsafe {
-            let ptr = self.as_ptr();
-            core::slice::from_raw_parts(ptr.add(self.begin), self.end.value() - self.begin)
-        }
-    }
-
-    #[inline]
-    pub const fn as_mut_slice(&mut self) -> &mut [T] {
-        // SAFETY: see above
-        unsafe {
-            let ptr = self.as_mut_ptr();
-            core::slice::from_raw_parts_mut(ptr.add(self.begin), self.end.value() - self.begin)
-        }
-    }
-}
-
-impl<T, const N: usize> Iterator for IntoIter<T, N> {
-    type Item = T;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.begin == self.end.value() {
-            None
-        } else {
-            // SAFETY: see above
-            unsafe {
-                let ptr = self.as_mut_ptr();
-                let value = ptr.add(self.begin).read();
-                self.begin += 1;
-                Some(value)
-            }
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let size = self.end.value() - self.begin;
-        (size, Some(size))
-    }
-}
-
-impl<T, const N: usize> DoubleEndedIterator for IntoIter<T, N> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        let mut end = self.end.value();
-        if self.begin == end {
-            None
-        } else {
-            // SAFETY: see above
-            unsafe {
-                let ptr = self.as_mut_ptr();
-                let on_heap = self.end.on_heap();
-                end -= 1;
-                self.end = TaggedLen::new(end, on_heap);
-                let value = ptr.add(end).read();
-                Some(value)
-            }
-        }
-    }
-}
-impl<T, const N: usize> ExactSizeIterator for IntoIter<T, N> {}
-impl<T, const N: usize> core::iter::FusedIterator for IntoIter<T, N> {}
 
 impl<T, const N: usize> SmallVec<T, N> {
     #[inline]
@@ -2090,29 +1622,6 @@ impl<T, const N: usize> Drop for SmallVec<T, N> {
     }
 }
 
-impl<T, const N: usize> Drop for IntoIter<T, N> {
-    fn drop(&mut self) {
-        // SAFETY: see above
-        unsafe {
-            let on_heap = self.end.on_heap();
-            let begin = self.begin;
-            let end = self.end.value();
-            let ptr = self.as_mut_ptr();
-            let _drop_dealloc = if on_heap {
-                let capacity = self.raw.heap.1;
-                Some(DropDealloc {
-                    ptr: NonNull::new_unchecked(ptr as *mut u8),
-                    size_bytes: capacity * size_of::<T>(),
-                    align: align_of::<T>(),
-                })
-            } else {
-                None
-            };
-            core::ptr::slice_from_raw_parts_mut(ptr.add(begin), end - begin).drop_in_place();
-        }
-    }
-}
-
 impl<T, const N: usize> core::ops::Deref for SmallVec<T, N> {
     type Target = [T];
 
@@ -2715,13 +2224,6 @@ impl<T: Clone, const N: usize> Clone for SmallVec<T, N> {
     }
 }
 
-impl<T: Clone, const N: usize> Clone for IntoIter<T, N> {
-    #[inline]
-    fn clone(&self) -> IntoIter<T, N> {
-        SmallVec::from(self.as_slice()).into_iter()
-    }
-}
-
 impl<T, const N: usize> Extend<T> for SmallVec<T, N> {
     #[inline]
     fn extend<I: IntoIterator<Item = T>>(&mut self, iter: I) {
@@ -2942,18 +2444,6 @@ impl<T, const N: usize> AsMut<[T]> for SmallVec<T, N> {
 impl<T: Debug, const N: usize> Debug for SmallVec<T, N> {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         f.debug_list().entries(self.iter()).finish()
-    }
-}
-
-impl<T: Debug, const N: usize> Debug for IntoIter<T, N> {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.debug_tuple("IntoIter").field(&self.as_slice()).finish()
-    }
-}
-
-impl<T: Debug, const N: usize> Debug for Drain<'_, T, N> {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        f.debug_tuple("Drain").field(&self.iter.as_slice()).finish()
     }
 }
